# Improve performance of color distance calculations by kernel fusion