@grlee77 2025-01-10
Not sure that there is a common use case for actually running these on full images or if it is most commonly used just to compare the perceptual difference of single pairs of colors. If the later, then the CPU will be better for running that case. 

I think perhaps there could be use cases like comparing perceptual differences among all pairs of colors in a categorical colormap, etc.